backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
groq==0.8.0
google-generativeai==0.5.4
weaviate-client==4.6.1
sentence-transformers==2.7.0
python-multipart==0.0.9
starlette==0.37.2
